Move _discern_referenced_rows to own source file
