+++ hstrat/test_drive/perfect_tracking/_GarbageCollectingPhyloTracker.py
@@ -4,44 +4,8 @@ import typing
 import numpy as np
 import pandas as pd
 
-from ..._auxiliary_lib import (
-    apply_swaps,
-    count_unique,
-    jit,
-    numba_bool_or_fallback,
-)
-
-# must be declared outside jit'ed function or numba fails
-_bool_t = numba_bool_or_fallback()
-
-# implemented as free function (not member) so self param doesn't interfere
-# with nopython directive
-@jit(nopython=True)
-def _discern_referenced_rows(
-    parentage_buffer: np.array,
-    num_records: int,
-    population_size: int,
-    below_row: int = 0,
-) -> np.array:
-
-    assert below_row >= 0
-    assert num_records >= below_row
-
-    referenced_rows = np.zeros(num_records - below_row, dtype=_bool_t)
-    for pop_position in range(population_size):
-        idx = num_records - population_size + pop_position
-        while (
-            idx != parentage_buffer[idx]
-            and idx >= below_row
-            and not referenced_rows[idx - below_row]
-        ):
-            referenced_rows[idx - below_row] = True
-            idx = parentage_buffer[idx]
-
-        if idx >= below_row:
-            referenced_rows[idx - below_row] = True
-
-    return referenced_rows
+from ..._auxiliary_lib import apply_swaps, count_unique
+from ._GarbageCollectingPhyloTracker_ import _discern_referenced_rows
 
 
 class GarbageCollectingPhyloTracker:
